Needs more bits

+++ lib/calendar.c
@@ -190,11 +190,11 @@ void free_week_boundaries(t_week_boundary boundary) {
 
 #ifdef EPD
 
-void draw_event(char *text, uint8_t quarter_offset, uint8_t quarter_length, uint8_t day_of_week) {
-    uint8_t x = 67 + day_of_week * 82;
-    uint8_t w = 72;
-    uint8_t y = 104 + (quarter_offset / 2) * 26 + (quarter_offset % 2) * 12;
-    uint8_t h = quarter_length * 11 + ((quarter_offset % 2) ? ((quarter_length / 2) * 1 + ((quarter_length - 1) / 2) * 3) : ((quarter_length / 2) * 3 + ((quarter_length - 1) / 2) * 1));
+void draw_event(char *text, uint16_t quarter_offset, uint16_t quarter_length, uint16_t day_of_week) {
+    uint16_t x = 67 + day_of_week * 82;
+    uint16_t w = 72;
+    uint16_t y = 104 + (quarter_offset / 2) * 26 + (quarter_offset % 2) * 12;
+    uint16_t h = quarter_length * 11 + ((quarter_offset % 2) ? ((quarter_length / 2) * 1 + ((quarter_length - 1) / 2) * 3) : ((quarter_length / 2) * 3 + ((quarter_length - 1) / 2) * 1));
 
     printf("x:%d, y:%d, w:%d, h:%d\n", x, y, w, h);
     Paint_DrawRectangle(x, y, x + w, y + h, RED, DOT_PIXEL_1X1, DRAW_FILL_FULL);
